added utils file for depth mesh utils
